修复double和int没搞清楚导致的逆天错误

+++ superstitioApi/java/superstitioapi/actions/DamageAllEnemiesAction.kt
@@ -59,7 +59,7 @@ class DamageAllEnemiesAction protected constructor(
             {
                 val monsters = AbstractDungeon.getCurrRoom().monsters.monsters
                 val targetsDamagesMap: MutableMap<AbstractCreature, Int> = HashMap()
-                for (i in 0 until min(monsters.size.toDouble(), multiDamages.size.toDouble()).toInt())
+                for (i in 0 until min(monsters.size, multiDamages.size))
                 {
                     targetsDamagesMap[monsters[i]] = multiDamages[i]
                 }
